pyproject.toml
[project]
name = "micro-internship-2025-winter"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]
